Move assertUtils file to utilities directory.
