+++ bftengine/tests/simpleStorage/main.cpp
@@ -12,7 +12,7 @@
 // file.
 
 #include "FileStorage.hpp"
-#include "../../src/bftengine/assertUtils.hpp"
+#include "assertUtils.hpp"
 
 #include <string.h>
 #include <cassert>
